Move device to gs.ts
