+++ src/devices/heiman.ts
@@ -371,30 +371,6 @@ const definitions: Definition[] = [
         exposes: [e.temperature(), e.humidity(), e.battery()],
     },
     {
-        zigbeeModel: ['SKHMP30-I1'],
-        model: 'SKHMP30-I1',
-        description: 'Smart metering plug',
-        vendor: 'GS',
-        fromZigbee: [fz.on_off, fz.electrical_measurement, fz.metering],
-        exposes: [e.switch(), e.power(), e.current(), e.voltage(), e.energy()],
-        toZigbee: [tz.on_off],
-        configure: async (device, coordinatorEndpoint, logger) => {
-            const endpoint = device.getEndpoint(1);
-            await reporting.bind(endpoint, coordinatorEndpoint, ['genOnOff', 'haElectricalMeasurement', 'seMetering']);
-            await reporting.onOff(endpoint);
-            await reporting.rmsVoltage(endpoint);
-            await reporting.rmsCurrent(endpoint);
-            await reporting.activePower(endpoint);
-            await reporting.readMeteringMultiplierDivisor(endpoint);
-            await reporting.currentSummDelivered(endpoint);
-            endpoint.saveClusterAttributeKeyValue('haElectricalMeasurement', {
-                acVoltageMultiplier: 1, acVoltageDivisor: 100,
-                acCurrentMultiplier: 1, acCurrentDivisor: 100,
-                acPowerMultiplier: 1, acPowerDivisor: 10,
-            });
-        },
-    },
-    {
         zigbeeModel: ['E_Socket'],
         model: 'HS2ESK-E',
         vendor: 'HEIMAN',
